Fixed query, authorization error for membership status change, and fixed format for lat and lng from string to num

+++ backend/db/models/venue.js
@@ -54,10 +54,10 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     lat: {
-      type: DataTypes.DECIMAL
+      type: DataTypes.FLOAT
     },
     lng: {
-      type: DataTypes.DECIMAL
+      type: DataTypes.FLOAT
     }
   }, {
     sequelize,
